change resource center to  authentication center
