+++ dlink-web/config/routes.ts
@@ -117,33 +117,33 @@ export default [
     ],
   },
   {
-    name: 'resourcecenter',
-    icon: 'CodeSandboxOutlined',
-    path: '/resourcecenter',
+    name: 'authenticationcenter',
+    icon: 'SafetyCertificateOutlined',
+    path: '/authenticationcenter',
     routes: [
       {
-        path: '/resourcecenter/usermanager',
+        path: '/authenticationcenter/usermanager',
         name: 'usermanager',
         icon: 'UserOutlined',
-        component: './ResourceCenter/UserManager',
+        component: './AuthenticationCenter/UserManager',
       },
       {
-        path: '/resourcecenter/rolemanager',
+        path: '/authenticationcenter/rolemanager',
         name: 'rolemanager',
         icon: 'TeamOutlined',
-        component: './ResourceCenter/RoleManager',
+        component: './AuthenticationCenter/RoleManager',
       },
       {
-        path: '/resourcecenter/namespacemanager',
+        path: '/authenticationcenter/namespacemanager',
         name: 'namespacemanager',
         icon: 'BulbOutlined',
-        component: './ResourceCenter/NamespaceManager',
+        component: './AuthenticationCenter/NamespaceManager',
       },
       {
-        path: '/resourcecenter/tenantmanager',
+        path: '/authenticationcenter/tenantmanager',
         name: 'tenantmanager',
         icon: 'SecurityScanOutlined',
-        component: './ResourceCenter/TenantManager',
+        component: './AuthenticationCenter/TenantManager',
       },
     ],
   },
