+++ dlink-web/src/locales/zh-CN/menu.ts
@@ -64,12 +64,11 @@ export default {
   'menu.registration.jar': 'Jar 管理',
   'menu.registration.document': '文档管理',
   'menu.registration.fragment': '全局变量管理',
-  'menu.registration.fragment': '全局变量管理',
-  'menu.resourcecenter': '资源中心',
-  'menu.resourcecenter.usermanager': '用户管理',
-  'menu.resourcecenter.namespacemanager': '命名空间管理',
-  'menu.resourcecenter.rolemanager': '角色管理',
-  'menu.resourcecenter.tenantmanager': '租户管理',
+  'menu.authenticationcenter': '认证中心',
+  'menu.authenticationcenter.usermanager': '用户管理',
+  'menu.authenticationcenter.namespacemanager': '命名空间管理',
+  'menu.authenticationcenter.rolemanager': '角色管理',
+  'menu.authenticationcenter.tenantmanager': '租户管理',
   'menu.settings': '配置中心',
   'menu.settings.flinkConfig': 'Flink 配置',
   'menu.dev.flink': 'Flink 计算框架',
