+++ dlink-web/src/pages/AuthenticationCenter/NamespaceManager/index.tsx
@@ -18,41 +18,34 @@
  */
 
 
-import React, {useEffect, useRef, useState} from "react";
+import React, {useRef, useState} from "react";
 import {DownOutlined, PlusOutlined} from '@ant-design/icons';
 import ProTable, {ActionType, ProColumns} from "@ant-design/pro-table";
 import {Button, Drawer, Dropdown, Menu, Modal} from 'antd';
 import {FooterToolbar, PageContainer} from '@ant-design/pro-layout';
 import ProDescriptions from '@ant-design/pro-descriptions';
-import {getStorageTenantId, handleAddOrUpdate, handleRemove, queryData} from "@/components/Common/crud";
-import {RoleTableListItem} from "@/pages/ResourceCenter/data.d";
-import RoleForm from "@/pages/ResourceCenter/RoleManager/components/RoleForm";
-import {getNameSpaceList} from "@/pages/ResourceCenter/service";
-import {connect} from "umi";
+import {handleAddOrUpdate, handleRemove, queryData, updateEnabled} from "@/components/Common/crud";
+import {NameSpaceTableListItem} from "@/pages/AuthenticationCenter/data.d";
+import NameSpaceForm from "@/pages/AuthenticationCenter/NamespaceManager/components/NameSpaceForm";
 
-const url = '/api/role';
+const url = '/api/namespace';
 
-const RoleFormList: React.FC<{}> = (props: any) => {
-  const {dispatch} = props;
-  const [row, setRow] = useState<RoleTableListItem>();
-  const [formValues, setFormValues] = useState<RoleTableListItem>();
-  const [modalVisible, handleModalVisible] = useState<boolean>(false);
+const NameSpaceFormList: React.FC<{}> = (props: any) => {
+  const [row, setRow] = useState<NameSpaceTableListItem>();
+  const [formValues, setFormValues] = useState<NameSpaceTableListItem>();
   const [updateModalVisible, handleUpdateModalVisible] = useState<boolean>(false);
+  const [modalVisible, handleModalVisible] = useState<boolean>(false);
   const actionRef = useRef<ActionType>();
-  const [selectedRowsState, setSelectedRows] = useState<RoleTableListItem[]>([]);
+  const [selectedRowsState, setSelectedRows] = useState<NameSpaceTableListItem[]>([]);
 
-  useEffect(() => {
-    getNameSpaceList(dispatch);
-  }, []);
-
-  const editAndDelete = (key: string | number, currentItem: RoleTableListItem) => {
+  const editAndDelete = (key: string | number, currentItem: NameSpaceTableListItem) => {
     if (key === 'edit') {
-      setFormValues(currentItem);
       handleUpdateModalVisible(true);
+      setFormValues(currentItem);
     } else if (key === 'delete') {
       Modal.confirm({
-        title: '删除角色',
-        content: '确定删除该角色吗？',
+        title: '删除命名空间',
+        content: '确定删除该命名空间吗？',
         okText: '确认',
         cancelText: '取消',
         onOk: async () => {
@@ -64,7 +57,7 @@ const RoleFormList: React.FC<{}> = (props: any) => {
   };
 
   const MoreBtn: React.FC<{
-    item: RoleTableListItem;
+    item: NameSpaceTableListItem;
   }> = ({item}) => (
     <Dropdown
       overlay={
@@ -80,54 +73,52 @@ const RoleFormList: React.FC<{}> = (props: any) => {
     </Dropdown>
   );
 
-  const columns: ProColumns<RoleTableListItem>[] = [
+  const columns: ProColumns<NameSpaceTableListItem>[] = [
     {
       title: 'ID',
       dataIndex: 'id',
       hideInTable: true,
       hideInSearch: true,
       sorter: true,
-    },
-    {
-      title: '角色编码',
-      dataIndex: 'roleCode',
       render: (dom, entity) => {
         return <a onClick={() => setRow(entity)}>{dom}</a>;
       },
     },
     {
-      title: '角色名称',
-      dataIndex: 'roleName',
+      title: '命名空间编码',
+      dataIndex: 'namespaceCode',
+      render: (dom, entity) => {
+        return <a onClick={() => setRow(entity)}>{dom}</a>;
+      },
     },
     {
       title: '所属租户',
       hideInSearch: true,
       render: (dom, entity) => {
-        return entity?.tenant?.tenantCode || '';
+        return entity.tenant.tenantCode;
       },
     },
-    // {
-    //   title: '是否删除',
-    //   dataIndex: 'isDelete',
-    //   hideInForm: true,
-    //   hideInSearch: true,
-    //   hideInTable: false,
-    //   filters: [
-    //     {
-    //       text: '未删除',
-    //       value: 0,
-    //     },
-    //     {
-    //       text: '已删除',
-    //       value: 1,
-    //     },
-    //   ],
-    //   filterMultiple: false,
-    //   valueEnum: {
-    //     true: {text: '已删除', status: 'Error'},
-    //     false: {text: '未删除', status: 'Success'},
-    //   },
-    // },
+    {
+      title: '是否启用',
+      dataIndex: 'enabled',
+      hideInTable: false,
+      hideInSearch: true,
+      filters: [
+        {
+          text: '已启用',
+          value: 1,
+        },
+        {
+          text: '未启用',
+          value: 0,
+        },
+      ],
+      filterMultiple: false,
+      valueEnum: {
+        true: {text: '已启用', status: 'Success'},
+        false: {text: '未启用', status: 'Error'},
+      },
+    },
     {
       title: '备注',
       dataIndex: 'note',
@@ -151,14 +142,14 @@ const RoleFormList: React.FC<{}> = (props: any) => {
       dataIndex: 'option',
       valueType: 'option',
       render: (_, record) => [
-        <Button type={"link"}
+        <a
           onClick={() => {
             handleUpdateModalVisible(true);
             setFormValues(record);
           }}
         >
           配置
-        </Button>,
+        </a>,
         <MoreBtn key="more" item={record}/>,
       ],
     },
@@ -166,8 +157,8 @@ const RoleFormList: React.FC<{}> = (props: any) => {
 
   return (
     <PageContainer>
-      <ProTable<RoleTableListItem>
-        headerTitle="角色管理"
+      <ProTable<NameSpaceTableListItem>
+        headerTitle="命名空间管理"
         actionRef={actionRef}
         rowKey="id"
         search={{
@@ -178,7 +169,7 @@ const RoleFormList: React.FC<{}> = (props: any) => {
             <PlusOutlined/> 新建
           </Button>,
         ]}
-        request={(params, sorter, filter) => queryData(url, {tenantId: getStorageTenantId(), sorter, filter})}
+        request={(params, sorter, filter) => queryData(url, {...params, sorter, filter})}
         columns={columns}
         rowSelection={{
           onChange: (_, selectedRows) => setSelectedRows(selectedRows),
@@ -190,7 +181,7 @@ const RoleFormList: React.FC<{}> = (props: any) => {
             <div>
               已选择 <a style={{fontWeight: 600}}>{selectedRowsState.length}</a> 项&nbsp;&nbsp;
               <span>
-  被删除的角色共 {selectedRowsState.length - selectedRowsState.reduce((pre, item) => pre + (item.isDelete ? 0 : 1), 0)} 个
+  被禁用的命名空间共 {selectedRowsState.length - selectedRowsState.reduce((pre, item) => pre + (item.enabled ? 1 : 0), 0)} 个
   </span>
             </div>
           }
@@ -198,8 +189,8 @@ const RoleFormList: React.FC<{}> = (props: any) => {
           <Button type="primary" danger
                   onClick={() => {
                     Modal.confirm({
-                      title: '删除角色',
-                      content: '确定删除选中的角色吗？',
+                      title: '删除命名空间',
+                      content: '确定删除选中的命名空间吗？',
                       okText: '确认',
                       cancelText: '取消',
                       onOk: async () => {
@@ -212,9 +203,39 @@ const RoleFormList: React.FC<{}> = (props: any) => {
           >
             批量删除
           </Button>
+          <Button type="primary"
+                  onClick={() => {
+                    Modal.confirm({
+                      title: '启用命名空间',
+                      content: '确定启用选中的命名空间吗？',
+                      okText: '确认',
+                      cancelText: '取消',
+                      onOk: async () => {
+                        await updateEnabled(url, selectedRowsState, true);
+                        setSelectedRows([]);
+                        actionRef.current?.reloadAndRest?.();
+                      }
+                    });
+                  }}
+          >批量启用</Button>
+          <Button danger
+                  onClick={() => {
+                    Modal.confirm({
+                      title: '禁用命名空间',
+                      content: '确定禁用选中的命名空间吗？',
+                      okText: '确认',
+                      cancelText: '取消',
+                      onOk: async () => {
+                        await updateEnabled(url, selectedRowsState, false);
+                        setSelectedRows([]);
+                        actionRef.current?.reloadAndRest?.();
+                      }
+                    });
+                  }}
+          >批量禁用</Button>
         </FooterToolbar>
       )}
-      <RoleForm
+      <NameSpaceForm
         onSubmit={async (value) => {
           const success = await handleAddOrUpdate(url, value);
           if (success) {
@@ -233,7 +254,7 @@ const RoleFormList: React.FC<{}> = (props: any) => {
       />
       {
         formValues && Object.keys(formValues).length ? (
-          <RoleForm
+          <NameSpaceForm
             onSubmit={async (value) => {
               const success = await handleAddOrUpdate(url, value);
               if (success) {
@@ -262,9 +283,9 @@ const RoleFormList: React.FC<{}> = (props: any) => {
         closable={false}
       >
         {row?.id && (
-          <ProDescriptions<RoleTableListItem>
+          <ProDescriptions<NameSpaceTableListItem>
             column={2}
-            title={row?.roleName}
+            title={row?.namespaceCode}
             request={async () => ({
               data: row || {},
             })}
@@ -279,4 +300,4 @@ const RoleFormList: React.FC<{}> = (props: any) => {
   );
 };
 
-export default connect()(RoleFormList);
+export default NameSpaceFormList;
