+++ dlink-web/src/pages/AuthenticationCenter/UserManager/components/PasswordForm.tsx
@@ -20,7 +20,7 @@
 
 import React, {useState} from 'react';
 import {Button, Form, Input, Modal} from 'antd';
-import {PasswordItem} from "@/pages/ResourceCenter/data.d";
+import {PasswordItem} from "@/pages/AuthenticationCenter/data.d";
 
 export type PasswordFormProps = {
   onCancel: (flag?: boolean) => void;
