+++ dlink-web/src/pages/AuthenticationCenter/UserManager/components/UserForm.tsx
@@ -20,14 +20,13 @@
 
 import React, {useState} from 'react';
 import {Button, Form, Input, Modal, Switch} from 'antd';
-import {NameSpaceTableListItem} from "@/pages/ResourceCenter/data.d";
-import {getStorageTenantId} from "@/components/Common/crud";
+import {UserTableListItem} from "@/pages/AuthenticationCenter/data.d";
 
-export type TenantFormProps = {
+export type UserFormProps = {
   onCancel: (flag?: boolean) => void;
-  onSubmit: (values: Partial<NameSpaceTableListItem>) => void;
+  onSubmit: (values: Partial<UserTableListItem>) => void;
   modalVisible: boolean;
-  values: Partial<NameSpaceTableListItem>;
+  values: Partial<UserTableListItem>;
 };
 
 const formLayout = {
@@ -35,17 +34,18 @@ const formLayout = {
   wrapperCol: {span: 13},
 };
 
-const NameSpaceForm: React.FC<TenantFormProps> = (props) => {
+const UserForm: React.FC<UserFormProps> = (props) => {
 
   const [form] = Form.useForm();
-  const [formVals, setFormVals] = useState<Partial<NameSpaceTableListItem>>({
-    id: props.values?.id,
-    tenantId: props.values?.tenantId,
-    namespaceCode: props.values?.namespaceCode,
-    enabled: props.values?.enabled,
-    note: props.values?.note,
-    createTime: props.values?.createTime,
-    updateTime: props.values?.updateTime,
+  const [formVals, setFormVals] = useState<Partial<UserTableListItem>>({
+    id: props.values.id,
+    username: props.values.username,
+    nickname: props.values.nickname,
+    password: props.values.password,
+    worknum: props.values.worknum,
+    mobile: props.values.mobile,
+    avatar: props.values.avatar,
+    enabled: props.values.enabled,
   });
 
   const {
@@ -54,41 +54,45 @@ const NameSpaceForm: React.FC<TenantFormProps> = (props) => {
     modalVisible,
   } = props;
 
+
   const submitForm = async () => {
     const fieldsValue = await form.validateFields();
-    fieldsValue.id = formVals.id;
-    setFormVals({...formVals,...fieldsValue});
-    handleSubmit({...formVals,...fieldsValue});
+    setFormVals({ ...formVals, ...fieldsValue });
+    handleSubmit({ ...formVals, ...fieldsValue });
   };
 
-  const renderContent = (formValsPara: Partial<NameSpaceTableListItem>) => {
+  const renderContent = (formVals: Partial<UserTableListItem>) => {
     return (
       <>
         <Form.Item
-          name="namespaceCode"
-          label="命名空间编号"
-          rules={[{required: true, message: '请输入命名空间唯一编码！'}]}>
-          <Input placeholder="请输入命名空间唯一编码"/>
+          name="username"
+          label="用户名"
+          rules={[{required: true, message: '请输入用户名！'}]}>
+          <Input placeholder="请输入唯一用户名"/>
         </Form.Item>
         <Form.Item
-          hidden={true}
-          name="tenantId"
-          label="所属租户"
-          >
-          <Input disabled defaultValue={getStorageTenantId()}/>
-        </Form.Item>
-        <Form.Item
-          name="note"
-          label="注释"
+          name="nickname"
+          label="昵称"
         >
-          <Input.TextArea placeholder="请输入描述信息" allowClear
-                          autoSize={{minRows: 3, maxRows: 10}}/>
+          <Input placeholder="请输入昵称"/>
+        </Form.Item>
+        <Form.Item
+          name="worknum"
+          label="工号"
+        >
+          <Input placeholder="请输入工号"/>
+        </Form.Item>
+        <Form.Item
+          name="mobile"
+          label="手机号"
+        >
+          <Input placeholder="请输入手机号"/>
         </Form.Item>
         <Form.Item
           name="enabled"
           label="是否启用">
           <Switch checkedChildren="启用" unCheckedChildren="禁用"
-                  defaultChecked={formValsPara.enabled}/>
+                  defaultChecked={formVals.enabled}/>
         </Form.Item>
       </>
     );
@@ -107,10 +111,10 @@ const NameSpaceForm: React.FC<TenantFormProps> = (props) => {
 
   return (
     <Modal
-      width={640}
+      width={1200}
       bodyStyle={{padding: '32px 40px 48px'}}
       destroyOnClose
-      title={formVals.id ? "修改命名空间" : "创建命名空间"}
+      title={formVals.id?"维护用户":"创建用户"}
       visible={modalVisible}
       footer={renderFooter()}
       onCancel={() => handleModalVisible()}
@@ -126,4 +130,4 @@ const NameSpaceForm: React.FC<TenantFormProps> = (props) => {
   );
 };
 
-export default NameSpaceForm;
+export default UserForm;
